[SYCL][ESIMD][NFC] Split up memory_properties.cpp
